Fix compilation on 32-bit archs (#465)
src/github.com/dgraph-io/ristretto/sketch_test.go:67:8: cannot use 12 << 30 (untyped int constant 12884901888) as int value in assignment (overflows)
src/github.com/dgraph-io/ristretto/store_test.go:209:23: cannot use 4340958203495 (untyped int constant) as int value in argument to z.KeyToHash (overflows)

+++ sketch_test.go
@@ -64,7 +64,7 @@ func TestSketchClear(t *testing.T) {
 }
 
 func TestNext2Power(t *testing.T) {
-	sz := 12 << 30
+	sz := int64(12) << 30
 	szf := float64(sz) * 0.01
 	val := int64(szf)
 	t.Logf("szf = %.2f val = %d\n", szf, val)
